+++ docs/handbook/sources/build.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ##
 ##   Documentation build script
-##   Last updated on May 11, 2026
+##   Last updated on May 23, 2026
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   SPDX-FileCopyrightText: 1998-2026 Paulo Moura <pmoura@logtalk.org>
@@ -167,6 +167,16 @@ sed '1,19d' ../../../library/hodge_rank/NOTES.md | pandoc -f gfm -t rst -o libra
 sed '1,19d' ../../../library/hook_flows/NOTES.md | pandoc -f gfm -t rst -o libraries/hook_flows.rst
 sed '1,19d' ../../../library/hook_objects/NOTES.md | pandoc -f gfm -t rst -o libraries/hook_objects.rst
 sed '1,19d' ../../../library/html/NOTES.md | pandoc -f gfm -t rst -o libraries/html.rst
+sed '1,19d' ../../../library/http/NOTES.md | pandoc -f gfm -t rst -o libraries/http.rst
+sed '1,19d' ../../../library/http_client/NOTES.md | pandoc -f gfm -t rst -o libraries/http_client.rst
+sed '1,19d' ../../../library/http_cookies/NOTES.md | pandoc -f gfm -t rst -o libraries/http_cookies.rst
+sed '1,19d' ../../../library/http_multipart/NOTES.md | pandoc -f gfm -t rst -o libraries/http_multipart.rst
+sed '1,19d' ../../../library/http_router/NOTES.md | pandoc -f gfm -t rst -o libraries/http_router.rst
+sed '1,19d' ../../../library/http_server/NOTES.md | pandoc -f gfm -t rst -o libraries/http_server.rst
+sed '1,19d' ../../../library/http_socket/NOTES.md | pandoc -f gfm -t rst -o libraries/http_socket.rst
+sed '1,19d' ../../../library/http_websocket/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket.rst
+sed '1,19d' ../../../library/http_websocket_messages/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket_messages.rst
+sed '1,19d' ../../../library/http_websocket_session/NOTES.md | pandoc -f gfm -t rst -o libraries/http_websocket_session.rst
 sed '1,19d' ../../../library/ica_projection/NOTES.md | pandoc -f gfm -t rst -o libraries/ica_projection.rst
 sed '1,19d' ../../../library/ids/NOTES.md | pandoc -f gfm -t rst -o libraries/ids.rst
 sed '1,19d' ../../../library/ieee_754/NOTES.md | pandoc -f gfm -t rst -o libraries/ieee_754.rst
@@ -226,6 +236,7 @@ sed '1,19d' ../../../library/nearest_centroid_classifier/NOTES.md | pandoc -f gf
 sed '1,19d' ../../../library/nested_dictionaries/NOTES.md | pandoc -f gfm -t rst -o libraries/nested_dictionaries.rst
 sed '1,19d' ../../../library/nmea/NOTES.md | pandoc -f gfm -t rst -o libraries/nmea.rst
 sed '1,19d' ../../../library/nmf_projection/NOTES.md | pandoc -f gfm -t rst -o libraries/nmf_projection.rst
+sed '1,19d' ../../../library/open_api/NOTES.md | pandoc -f gfm -t rst -o libraries/open_api.rst
 sed '1,19d' ../../../library/optics_clusterer/NOTES.md | pandoc -f gfm -t rst -o libraries/optics_clusterer.rst
 sed '1,19d' ../../../library/optionals/NOTES.md | pandoc -f gfm -t rst -o libraries/optionals.rst
 sed '1,19d' ../../../library/options/NOTES.md | pandoc -f gfm -t rst -o libraries/options.rst