+++ paths/paths.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Default library paths
-%  Last updated on May 17, 2026
+%  Last updated on May 23, 2026
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  SPDX-FileCopyrightText: 1998-2026 Paulo Moura <pmoura@logtalk.org>
@@ -199,8 +199,13 @@ logtalk_library_path(html, library('html/')).
 logtalk_library_path(http, library('http/')).
 logtalk_library_path(http_client, library('http_client/')).
 logtalk_library_path(http_cookies, library('http_cookies/')).
+logtalk_library_path(http_multipart, library('http_multipart/')).
+logtalk_library_path(http_router, library('http_router/')).
 logtalk_library_path(http_server, library('http_server/')).
 logtalk_library_path(http_socket, library('http_socket/')).
+logtalk_library_path(http_websocket, library('http_websocket/')).
+logtalk_library_path(http_websocket_messages, library('http_websocket_messages/')).
+logtalk_library_path(http_websocket_session, library('http_websocket_session/')).
 logtalk_library_path(ica_projection, library('ica_projection/')).
 logtalk_library_path(ica_projection, library('ica_projection/')).
 logtalk_library_path(ids, library('ids/')).